api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "second-brain-api"
version = "0.1.0"
description = "Second Brain ingest API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "pyyaml",
    "httpx",
    "apify-client",
    "openai",
    "readability-lxml",
    "beautifulsoup4",
    "lxml",
    "pytest",
    "pytest-asyncio",
]

[project.optional-dependencies]
dev = []

[tool.hatch.build.targets.wheel]
packages = ["src/brain_api"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
